requirements.txt
asgiref
certifi
chardet
defusedxml
dj-database-url
Django==4.2.7
django-allauth==0.57.0
django-ckeditor==6.7.0
django-js-asset==2.1.0
django-multiselectfield==0.1.12
gunicorn
idna
oauthlib
Pillow==10.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python3-openid
pytz
requests
requests-oauthlib
sqlparse
urllib3
whitenoise
huggingface_hub==0.19.4
transformers==4.36.2
python-dotenv==1.0.0
SpeechRecognition==3.10.0
google-generativeai==0.3.2
langchain==0.1.5
faiss-cpu==1.7.4
sentence-transformers==2.3.1
